+++ lib/frp/src/node.rs
@@ -1,7 +1,9 @@
 pub mod class;
 pub mod id;
 pub mod label;
+pub mod wrapper;
 
 pub use class::*;
 pub use id::*;
 pub use label::*;
+pub use wrapper::*;
